contiunue training 96

+++ train/ToonNetTrainer_cont96.py
@@ -206,7 +206,7 @@ class ToonNetTrainer:
 
     def learning_rate_cont(self):
         # Define learning rate schedule
-        boundaries = [900000, 1000000, 1100000]
+        boundaries = [np.int64(900000), np.int64(1000000), np.int64(1100000)]
         values = [0.0002, 0.0001, 0.00005, 0.000025]
         return tf.train.piecewise_constant(self.global_step, boundaries=boundaries, values=values)
